// apps/inatrace-fe/src/api/model/apiProcessingEvidenceType.ts
export interface ApiProcessingEvidenceType {
  /**
   * Entity id
   */
  id?: number;
  /**
   * code
   */
  code: string;
  /**
   * label
   */
  label: string;
  /**
   * type of evidence
   */
  type?: ApiProcessingEvidenceType.TypeEnum;
  /**
   * if evidence is of fairness type
   */
  fairness?: boolean;
  /**
   * if evidence is of provenance type
   */
  provenance?: boolean;
  /**
   * if evidence is of quality type
   */
  quality?: boolean;
  /**
   * whether the evidence is mandatory
   */
  mandatory?: boolean;
  /**
   * whether the evidence is required on quote
   */
  requiredOnQuote?: boolean;
  /**
   * a group in which at least one document has to be provided
   */
  requiredOneOfGroupIdForQuote?: string;
  /**
   * Translations for processing evidence type
   */
  translations?: Array<ApiProcessingEvidenceTypeTranslation>;
}

/**
 * Namespace for property- and property-value-enumerations of ApiProcessingEvidenceType.
 */
export namespace ApiProcessingEvidenceType {
// ...
  export enum TypeEnum {
    DOCUMENT = 'DOCUMENT',
    FIELD = 'FIELD',
    CALCULATED = 'CALCULATED',
  }

  export function formMetadata() {
    return {
      metadata: formMetadata,
      classname: 'ApiProcessingEvidenceType',
      vars: [
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'id',
          classname: 'ApiProcessingEvidenceType',
          dataType: 'number',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
// ...
      validators: {
        id: [],
        code: [['required']],
        label: [['required']],
        type: [],
        fairness: [],
        provenance: [],
        quality: [],
        mandatory: [],
        requiredOnQuote: [],
        requiredOneOfGroupIdForQuote: [],
        translations: [],
      },
    };
  }
// ...
}